Fix confread.c parser to not fail on nexthop= settings with ipv6.
Moved the proper configurations in ipsec.conf.common, which now loads
with these changes for the 6in4 and 4in6 cases.

This issue is a little tricky. Currently, the parser assumes that
the family for left,leftnexthop,leftsubnet is always the same. This
is false for 6in4 and 4in6 tunnels.

I've made the parser more lenient but it might now allow really bogus
combinations of left=1.2.3.4 and leftnexthop=::1

The problem is that the parser starts in validate_end() with pulling
the connaddrfamily value and base all checks off of that. What we will
need to do is determine the family of (1) left/right/leftnexthop/rightnexthop
and (2) leftsubnet/rightsubnet and see if the two sets are consistent.

This is somewhat hard due to %defaultroute, %any and vhost/vnet options.

connaddrfamily= as an option is horrible when configuring 4in6 and 6in4.
Currently, best results are to keep the connaddrfamily= to the family
of the subnet, and in absense of that, of the left/right